feat(combat): implement asynchronous turn resolution and tactical visualizations
- Implement the Observer pattern for UI state management between Deployment and Combat phases.
- Create ActionArrowController to render movement and attack vectors based on stats (Shift+RMB for ranged attacks).
- Add CombatOrder data structure to queue player-issued commands.
- Refactor BattleManager with asynchronous turn resolution (Coroutines) for smooth token movement on the board.
- ntegrate the command loop with CombatEngine and auto-sync unit UI stats post-combat (UpdateVisual).

+++ Assets/Scripts/Engine/CombatEngine.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class CombatEngine
 {
-    public void ResolveEngagement(Unit attacker, Unit defender, EngagementType engagement, int distance)
+    public void ResolveEngagement(Unit attacker, Unit defender, EngagementType engagement, float distance)
     {
         int totalDamageToDefender = 0;
         int totalDamageToAttacker = 0;
